refactor: split TelegramChannel.js into TelegramBot + ConsoleHandler
- Extract TelegramBot class to TelegramBot.js (734 LOC)
- Extract console mode logic to ConsoleHandler.js (63 LOC)
- TelegramChannel.js reduced from 1133 → 354 LOC (manager only)
- Re-export TelegramBot from TelegramChannel for backwards compat
- Completes roadmap item 4.2
